Add files via upload
We will use real-life datasets to demonstrate the basics of machine learning, which
include preprocessing data for machine learning models and building a linear
regression model.
We will use a Car Price dataset
These data are the results of an automobile consulting company to understand the
factors on which the pricing of cars depends. The analysis determined the quantities of 26
constituents, found in each of the 23 types of cars.